Added new component for displaying MostActiveRules
